feat: refactor Blog module to use BlogPost as aggregate root

+++ src/Modules/Blog/Explorer.Blog.Infrastructure/Database/Repositories/BlogDbRepository.cs
@@ -25,7 +25,7 @@ namespace Explorer.Blog.Infrastructure.Database.Repositories
             return blog;
         }
 
-        public BlogPost Get(int id)
+        public BlogPost Get(long id)
         {
             var entity = _dbSet.Find(id);
             if (entity == null) throw new NotFoundException("Blog not found: " + id);
@@ -51,7 +51,7 @@ namespace Explorer.Blog.Infrastructure.Database.Repositories
             return blog;
         }
 
-        public void Delete(int id)
+        public void Delete(long id)
         {
             var entity = Get(id);
             _dbSet.Remove(entity);
